requirements.txt
# Core dependencies - Production only
streamlit>=1.31.0
face-recognition>=1.3.0
opencv-python-headless>=4.9.0.80
numpy>=1.24.3,<2.0.0
Pillow>=10.2.0

# Authentication and Database
bcrypt>=4.1.2
psycopg2-binary>=2.9.9
